+++ extras/AFC_BoxTurtle.py
@@ -83,9 +83,12 @@ class afcBoxTurtle(afcUnit):
                             msg +="<span class=primary--text> in ToolHead</span>"
                             if CUR_LANE.extruder_obj.tool_start == "buffer":
                                 msg += "<span class=warning--text>\n Ram sensor enabled, confirm tool is loaded</span>"
-                            self.AFC.SPOOL.set_active_spool(CUR_LANE.spool_id)
-                            self.AFC.FUNCTION.afc_led(CUR_LANE.led_tool_loaded, CUR_LANE.led_index)
-                            CUR_LANE.status = 'Tooled'
+
+                            if self.AFC.FUNCTION.get_current_lane() == CUR_LANE.name:
+                                self.AFC.SPOOL.set_active_spool(CUR_LANE.spool_id)
+                                self.AFC.FUNCTION.afc_led(CUR_LANE.led_tool_loaded, CUR_LANE.led_index)
+                                CUR_LANE.status = 'Tooled'
+
                             CUR_LANE.enable_buffer()
                         else:
                             if CUR_LANE.get_toolhead_sensor_state() == True: